restructure the menuitem components
